deprecated APIs fixes
