+++ src/main/java/com/crowdin/client/CrowdinProjectCacheProvider.java
@@ -18,6 +18,7 @@ import java.util.HashMap;
 import java.util.List;
 import java.util.Map;
 
+//TODO move to Project-level service
 public class CrowdinProjectCacheProvider {
 
     private static CrowdinProjectCache crowdinProjectCache;
