+++ src/main/java/com/crowdin/ui/panel/upload/action/RefreshAction.java
@@ -3,6 +3,7 @@ package com.crowdin.ui.panel.upload.action;
 import com.crowdin.action.ActionContext;
 import com.crowdin.action.BackgroundAction;
 import com.crowdin.client.FileBean;
+import com.crowdin.service.ProjectService;
 import com.crowdin.ui.panel.CrowdinPanelWindowFactory;
 import com.crowdin.ui.panel.upload.UploadWindow;
 import com.crowdin.util.FileUtil;
@@ -10,7 +11,6 @@ import com.crowdin.util.NotificationUtil;
 import com.intellij.icons.AllIcons;
 import com.intellij.openapi.actionSystem.AnActionEvent;
 import com.intellij.openapi.application.ApplicationManager;
-import com.intellij.openapi.components.ServiceManager;
 import com.intellij.openapi.progress.ProcessCanceledException;
 import com.intellij.openapi.progress.ProgressIndicator;
 import com.intellij.openapi.project.Project;
@@ -48,9 +48,7 @@ public class RefreshAction extends BackgroundAction {
         e.getPresentation().setEnabled(false);
         isInProgress.set(true);
         try {
-            UploadWindow window = ServiceManager
-                    .getService(project, CrowdinPanelWindowFactory.ProjectService.class)
-                    .getUploadWindow();
+            UploadWindow window = project.getService(ProjectService.class).getUploadWindow();
 
             if (window == null) {
                 return;
