+++ src/main/java/com/crowdin/util/NotificationUtil.java
@@ -1,8 +1,8 @@
 package com.crowdin.util;
 
 import com.intellij.notification.Notification;
-import com.intellij.notification.NotificationDisplayType;
 import com.intellij.notification.NotificationGroup;
+import com.intellij.notification.NotificationGroupManager;
 import com.intellij.notification.NotificationType;
 import com.intellij.notification.Notifications;
 import com.intellij.openapi.application.ApplicationManager;
@@ -17,12 +17,10 @@ import java.util.Date;
 public final class NotificationUtil {
 
     private static final NotificationGroup GROUP_DISPLAY_ID_INFO =
-            new NotificationGroup("Crowdin",
-                    NotificationDisplayType.BALLOON, true);
+            NotificationGroupManager.getInstance().getNotificationGroup("Crowdin BALLON");
 
     private static final NotificationGroup GROUP_DISPLAY_ID_INFO_LOG =
-        new NotificationGroup("Crowdin",
-            NotificationDisplayType.NONE, true);
+            NotificationGroupManager.getInstance().getNotificationGroup("Crowdin NONE");
     private static boolean isDebug = false;
     private static final SimpleDateFormat logDateFormat = new SimpleDateFormat("yyyy-MM-dd HH:mm:ss.SSS");
 
@@ -47,7 +45,7 @@ public final class NotificationUtil {
 
     private static void showMessage(Project project, String message, NotificationType type) {
         ApplicationManager.getApplication().invokeLater(() -> {
-            Notification notification = GROUP_DISPLAY_ID_INFO.createNotification(TITLE, message, type, null);
+            Notification notification = GROUP_DISPLAY_ID_INFO.createNotification(TITLE, message, type);
             Notifications.Bus.notify(notification, project);
         });
     }
@@ -71,7 +69,7 @@ public final class NotificationUtil {
         if (isDebug) {
             String formattedMessage = String.format("%s %s : %s", logDateFormat.format(new Date()), level, message);
             ApplicationManager.getApplication().invokeLater(() -> {
-                Notification notification = GROUP_DISPLAY_ID_INFO_LOG.createNotification(TITLE, formattedMessage, type, null);
+                Notification notification = GROUP_DISPLAY_ID_INFO_LOG.createNotification(TITLE, formattedMessage, type);
                 Notifications.Bus.notify(notification, project);
             });
         }
